Rename files and update imports
