+++ src/components/GitStatus.tsx
@@ -1,8 +1,7 @@
 import { useMemo } from "react";
 import { ActionPanel, List } from "@raycast/api";
 import { showFailureToast, useExec } from "@raycast/utils";
-import type { GitStatus } from "../utils/status.js";
-import { parseGitStatus } from "../utils/status.js";
+import { parseGitPorcelainStatus } from "../utils/porcelain.js";
 import { useRepoStorage } from "../hooks/useRepo.js";
 import { GitStatusItem } from "./GitStatus/GitStatusItem.js";
 import { RemoteGitActions } from "./GitStatus/RemoteGitActions.js";
@@ -20,7 +19,7 @@ export function GitStatus() {
     onError: (error) => {
       showFailureToast(error, { title: "Could not fetch git status" });
     },
-    parseOutput: ({ stdout }) => parseGitStatus(stdout),
+    parseOutput: ({ stdout }) => parseGitPorcelainStatus(stdout),
   });
 
   const showDetails = useMemo(() => !!repo.value && !!data?.files.length, [data?.files.length, repo.value]);
