+++ src/components/GitStatus/GitStatusItemDetail.tsx
@@ -1,6 +1,6 @@
 import { List } from "@raycast/api";
 import type { BranchInfo } from "../../utils/branch.js";
-import type { StatusInfo } from "../../utils/status.js";
+import type { StatusInfo } from "../../utils/porcelain.js";
 import { GitStatusTags } from "./GitStatusTags.js";
 import { useMemo } from "react";
