+++ src/utils/porcelain.ts
@@ -1,236 +1,102 @@
-import { SubmoduleStatus, parseSubmodule } from "./submodule.js";
+import { BranchInfo, parseBranchHeaders } from "./branch.js";
+import { parsePorcelainStatus, type XYStatus as GitStatus } from "./file.js";
+export type { GitStatus };
 
-export type XYStatus =
-  | /** unmodified */
-  "."
-  /** modified */
-  | "M"
-  /** file type changed (regular file, symbolic link or submodule) */
-  | "T"
-  /** added */
-  | "A"
-  /** deleted */
-  | "D"
-  /** renamed */
-  | "R"
-  /** copied (if config option status.renames is set to "copies") */
-  | "C"
-  /** updated but unmerged */
-  | "U"
-  /** untracked */
-  | "?"
-  /** ignored */
-  | "!";
+type GitStatusFormat = "changed" | "renamed" | "unmerged" | "untracked" | "ignored";
 
-type LineIndicator =
-  | /** Ordinary changed entry */
-  "1"
-  /** Renamed or copied entry */
-  | "2"
-  /** Unmerged entry */
-  | "u"
-  /** Untracked entry */
-  | "?"
-  /** Ignored entry */
-  | "!";
-
-interface LineFormat {
-  indicator: LineIndicator;
+export interface StatusInfo {
+  format: GitStatusFormat;
+  /** Path of the file */
+  fileName: string;
+  /** Original path of the file if it was renamed/copied */
+  origPath?: string;
   /** Status of the index */
-  indexChanges: XYStatus;
-  /** Status of the woking tree */
-  workingChanges: XYStatus;
-  /** Status of the git submodule */
-  submodule: SubmoduleStatus;
-  /** The octal file mode in HEAD */
-  mH: string;
-  /** The octal file mode in the index */
-  mI: string;
-  /** The octal file mode in the worktree */
-  mW: string;
-  /** The octal file mode in stage 1. */
-  m1: string;
-  /** The octal file mode in stage 2. */
-  m2: string;
-  /** The octal file mode in stage 3. */
-  m3: string;
-  /** The object name in HEAD */
-  hH: string;
-  /** The object name in the index */
-  hI: string;
-  /** The object name in stage 1. */
-  h1: string;
-  /** The object name in stage 2. */
-  h2: string;
-  /** The object name in stage 3. */
-  h3: string;
-  /** Rename or copy */
-  rc: "R" | "C";
-  /** Score denoting the percentage of similarity between the source and target of the move or copy */
-  rcScore: number;
-  /** The pathname. In a renamed/copied entry, this is the target path */
-  path: string;
-  /** The pathname in the commit at HEAD or in the index */
-  origPath: string;
+  staged: GitStatus;
+  /** Status of the working tree */
+  unstaged: GitStatus;
+  /** Is this a submodule */
+  submodule?: boolean;
 }
 
-interface ChangedFile
-  extends Pick<
-    LineFormat,
-    "indicator" | "indexChanges" | "workingChanges" | "submodule" | "mH" | "mI" | "mW" | "hH" | "hI" | "path"
-  > {
-  indicator: "1";
-}
-interface RenamedFile
-  extends Pick<
-    LineFormat,
-    | "indicator"
-    | "indexChanges"
-    | "workingChanges"
-    | "submodule"
-    | "mH"
-    | "mI"
-    | "mW"
-    | "hH"
-    | "hI"
-    | "rc"
-    | "rcScore"
-    | "path"
-    | "origPath"
-  > {
-  indicator: "2";
-}
-interface UnmergedFile
-  extends Pick<
-    LineFormat,
-    | "indicator"
-    | "indexChanges"
-    | "workingChanges"
-    | "submodule"
-    | "m1"
-    | "m2"
-    | "m3"
-    | "mW"
-    | "h1"
-    | "h2"
-    | "h3"
-    | "path"
-  > {
-  indicator: "u";
-}
-interface UntrackedFile extends Pick<LineFormat, "indicator" | "path"> {
-  indicator: "?";
-}
-interface IgnoredFile extends Pick<LineFormat, "indicator" | "path"> {
-  indicator: "!";
-}
-
-export type PorcelainStatus = ChangedFile | RenamedFile | UnmergedFile | UntrackedFile | IgnoredFile;
-
-function parseChanges(xy: string): [XYStatus, XYStatus] {
-  return [xy.charAt(0) as XYStatus, xy.charAt(1) as XYStatus];
-}
-
-/** Match unescaped space characters, in case there are spaces in a filename */
-const spaceRegex = /(?<!\\) /;
-
-function parsePaths(strings: string[]): [string, string] {
-  const rejoinedPaths = strings.join(" ");
-  // in the case of a rename pathnames are separated by a tab character
-  const [path, origPath] = rejoinedPaths.split(/\t/);
-  return [path, origPath];
-}
-
-function parseChangedFile(line: string): ChangedFile {
-  const [, xy, sub, mH, mI, mW, hH, hI, ...paths] = line.split(spaceRegex);
-  const [indexChanges, workingChanges] = parseChanges(xy);
-  const [path] = parsePaths(paths);
-
-  return {
-    indicator: "1",
-    indexChanges,
-    workingChanges,
-    submodule: parseSubmodule(sub),
-    mH,
-    mI,
-    mW,
-    hH,
-    hI,
-    path,
-  };
-}
-
-function parseRenamedFile(line: string): RenamedFile {
-  const [, xy, sub, mH, mI, mW, hH, hI, xScore, ...paths] = line.split(spaceRegex);
-  const [indexChanges, workingChanges] = parseChanges(xy);
-  const [rc, score] = xScore.split(/\s/) as ["R" | "C", string];
-  const [path, origPath] = parsePaths(paths);
-
-  return {
-    indicator: "2",
-    indexChanges,
-    workingChanges,
-    submodule: parseSubmodule(sub),
-    mH,
-    mI,
-    mW,
-    hH,
-    hI,
-    rc,
-    rcScore: +score,
-    path,
-    origPath,
-  };
-}
-
-function parseUnmergedFile(line: string): UnmergedFile {
-  const [, xy, sub, m1, m2, m3, mW, h1, h2, h3, ...paths] = line.split(spaceRegex);
-  const [indexChanges, workingChanges] = parseChanges(xy);
-  const [path] = parsePaths(paths);
-
-  return {
-    indicator: "u",
-    indexChanges,
-    workingChanges,
-    submodule: parseSubmodule(sub),
-    m1,
-    m2,
-    m3,
-    mW,
-    h1,
-    h2,
-    h3,
-    path,
-  };
-}
-
-export function parsePorcelainStatus(line: string): PorcelainStatus {
-  const indicator = line.charAt(0) as LineIndicator;
-
-  switch (indicator) {
+/**
+ * Git Status Porcelain Format Version 2 have 3 lines indicated by the first character:
+ * - `1`: A changed file
+ * - `2`: A renamed file
+ * - `u`: An unmerged file
+ *
+ * There are also two other options:
+ * - `?` An untracked file
+ * - `!` An ignored file
+ */
+function lineFormat(format: string): GitStatusFormat {
+  switch (format) {
     case "1":
-      return parseChangedFile(line);
+      return "changed";
     case "2":
-      return parseRenamedFile(line);
+      return "renamed";
     case "u":
-      return parseUnmergedFile(line);
-    case "?": {
-      const [, ...paths] = line.split(spaceRegex);
-      const [path] = parsePaths(paths);
-
-      return {
-        indicator: "?",
-        path,
-      };
-    }
-    case "!": {
-      const [, ...paths] = line.split(" ");
-      const [path] = parsePaths(paths);
-
-      return {
-        indicator: "!",
-        path,
-      };
-    }
+      return "unmerged";
+    case "?":
+      return "untracked";
+    default:
+      return "ignored";
   }
+}
+
+/**
+ * Parse a row from git status in the Porcelain 2 format
+ * @param dataRow
+ * @returns
+ */
+function parseGitFileStatus(dataRow: string): StatusInfo {
+  const fields = parsePorcelainStatus(dataRow);
+  if (fields.indicator === "?" || fields.indicator === "!") {
+    return {
+      format: "untracked",
+      fileName: fields.path,
+      staged: ".",
+      unstaged: fields.indicator,
+    };
+  }
+  const format = lineFormat(fields.indicator);
+
+  return {
+    format,
+    fileName: fields.path,
+    origPath: fields.indicator === "2" ? fields.origPath : undefined,
+    staged: fields.indexChanges,
+    unstaged: fields.workingChanges,
+    submodule: fields.submodule.isSubmodule,
+  };
+}
+
+export function parseGitPorcelainStatus(porcelainStatus: string):
+  | {
+      branch: BranchInfo;
+      files: StatusInfo[];
+    }
+  | undefined {
+  if (!porcelainStatus) {
+    return;
+  }
+
+  const status = porcelainStatus.split("\n");
+  const branch: BranchInfo = {
+    name: "",
+    commit: "",
+  };
+  const files: StatusInfo[] = [];
+  status.forEach((statusRow) => {
+    if (!statusRow) {
+      return;
+    }
+
+    if (statusRow.startsWith("#")) {
+      parseBranchHeaders(statusRow, branch);
+    } else {
+      files.push(parseGitFileStatus(statusRow));
+    }
+  });
+
+  return { branch, files };
 }
